Added 'get <file>' command to ftp client
Takes input 'get <file> <file> ...' altough only checks the first
two files.  First is file to get remotely.  If only that file is
provided, then a local file of that name is created or overwritten
and the remote file copied into it.

If two files are provided, the first is the remote file to be copied,
but the second is the local file to be created or overwritten (allowing
renaming of remote files) and copied into.

No in depth testing, just some simple manual tests.  Will complete
more testing before merging with master.